PD2 Season 2 update
In the cases where there were inconsistencies between the official patch notes and the in-game tooltips, I mostly went with the patch notes with the assumption that hotfixes will update them at some point.

All skills are updated, and the only major issue seems to be Fire Golem. It seems to operate with some kind of hidden base damage, as the synergy bonuses are nowhere near 20%. I left a big red message on the skill tooltip, and used 1% synergies since that seems to be very close to displayed damage values after the first couple levels.